Move some stuff around
This patch adds base pbr support for the main executable.
Also added tox support for testing pep8, and unit testing.
