+++ .stestr.conf
@@ -1,0 +1,3 @@
+[DEFAULT]
+test_path=${OS_TEST_PATH:-./openstack_exporter/tests}
+top_dir=./
