+++ openstack_exporter/collectors/CinderCollector.py
@@ -1,0 +1,30 @@
+# All Rights Reserved.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+
+from prometheus_client.core import GaugeMetricFamily
+
+from BaseCollector import BaseCollector
+
+
+class CinderCollector(BaseCollector):
+    def __init__(self):
+        pass
+
+    def describe(self):
+        # all your metrics go in here, without a label
+        yield GaugeMetricFamily('cinder_foo_stat', 'Cinder status of foo')
+
+    def collect(self):
+        # logic goes in here
+        pass
