+++ setup.cfg
@@ -1,0 +1,52 @@
+[bdist_wheel]
+universal = 1
+
+[metadata]
+name = openstack-exporter
+summary = "App to export metrics to prometheus from OpenStack services."
+description_file =
+  README.md
+author = Walter A. Boring IV
+author-email = waboring@hemna.com
+home-page = http://github.com/sapcc/openstack-exporter
+license_file = LICENSE
+classifier =
+    Environment :: OpenStack
+    Intended Audience :: Information Technology
+    Intended Audience :: System Administrators
+    License :: OSI Approved :: Apache Software License
+    Natural Language :: English
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+
+[global]
+setup-hooks =
+    pbr.hooks.setup_hook
+
+[files]
+packages =
+    openstack_exporter
+
+[entry_points]
+console_scripts =
+    openstack_exporter = openstack_exporter.exporter:main
+
+[egg_info]
+tag_build =
+tag_date = 0
+tag_svn_revision = 0
+
+[build_sphinx]
+warning-is-error = 1
+source-dir = doc/source
+build-dir = doc/build
+all_files = 1
+builders = html
+
+[pbr]
+skip_authors = 1
+
+[upload_sphinx]
+upload-dir = doc/build/html
