+++ test-requirements.txt
@@ -1,0 +1,14 @@
+# The order of packages is significant, because pip processes them in the order
+# of appearance. Changing the order has an impact on the overall integration
+# process, which may cause wedges in the gate later.
+# Hacking already pins down pep8, pyflakes and flake8
+
+hacking>=3.1.0,<3.2.0 # Apache-2.0
+coverage!=4.4,>=4.0 # Apache-2.0
+ddt>=1.0.1 # MIT
+fixtures>=3.0.0 # Apache-2.0/BSD
+reno>=3.1.0 # Apache-2.0
+requests-mock>=1.2.0 # Apache-2.0
+tempest>=17.1.0 # Apache-2.0
+testtools>=2.2.0 # MIT
+stestr>=1.0.0 # Apache-2.0
