+++ tox.ini
@@ -1,0 +1,90 @@
+[tox]
+distribute = False
+envlist = py36,py38,pep8
+minversion = 3.1.0
+skipsdist = True
+skip_missing_interpreters = true
+# this allows tox to infer the base python from the environment name
+# and override any basepython configured in this file
+ignore_basepython_conflict=true
+
+[testenv]
+basepython = python3
+usedevelop = True
+install_command = pip install {opts} {packages}
+setenv =
+    VIRTUAL_ENV={envdir}
+    OS_TEST_PATH=./openstack_exporter/tests
+    OS_STDOUT_CAPTURE=1
+    OS_STDERR_CAPTURE=1
+    OS_TEST_TIMEOUT=60
+passenv = *_proxy *_PROXY
+
+deps =
+       -c{env:UPPER_CONSTRAINTS_FILE:https://releases.openstack.org/constraints/upper/master}
+       -r{toxinidir}/requirements.txt
+       -r{toxinidir}/test-requirements.txt
+commands = find . -type f -name "*.pyc" -delete
+           stestr run {posargs}
+           stestr slowest
+whitelist_externals = find
+
+[testenv:pep8]
+commands = 
+  flake8 {posargs} .
+
+[testenv:pylint]
+deps =
+  -c{env:UPPER_CONSTRAINTS_FILE:https://releases.openstack.org/constraints/upper/master}
+  -r{toxinidir}/requirements.txt
+       pylint==1.9.1
+commands = bash tools/lintstack.sh
+whitelist_externals = bash
+
+[testenv:venv]
+commands = {posargs}
+
+[testenv:cover]
+setenv =
+    {[testenv]setenv}
+    PYTHON=coverage run --source openstack_exporter --parallel-mode
+commands =
+    stestr run {posargs}
+    coverage combine
+    coverage html -d cover
+    coverage xml -o cover/coverage.xml
+
+[testenv:docs]
+deps =
+  -c{env:UPPER_CONSTRAINTS_FILE:https://releases.openstack.org/constraints/upper/master}
+  -r{toxinidir}/requirements.txt
+  -r{toxinidir}/doc/requirements.txt
+commands = sphinx-build -W -b html doc/source doc/build/html
+
+[testenv:pdf-docs]
+deps =
+  {[testenv:docs]deps}
+commands =
+  {[testenv:docs]commands}
+  sphinx-build -W -b latex doc/source doc/build/pdf
+  make -C doc/build/pdf
+whitelist_externals =
+  make
+  cp
+
+[testenv:releasenotes]
+deps =
+  -c{env:UPPER_CONSTRAINTS_FILE:https://releases.openstack.org/constraints/upper/master}
+  -r{toxinidir}/requirements.txt
+  -r{toxinidir}/doc/requirements.txt
+commands = sphinx-build -a -E -W -d releasenotes/build/doctrees -b html releasenotes/source releasenotes/build/html
+
+# The OS_CACERT environment variable should be passed to the test
+# environments to specify a CA bundle file to use in verifying a
+# TLS (https) server certificate.
+passenv = OS_*
+
+[flake8]
+show-source = True
+ignore = H301,H404,H405,E122,E123,E128,E251,W504
+exclude=.venv,.git,.tox,dist,doc,*lib/python*,*egg,build
